# apps/server/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "ApiResponse_null_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "anyOf": [],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
...
      "CastData": {
        "properties": {
          "name": {
            "type": "string"
          },
          "character": {
            "type": "string"
          },
          "profileImage": {
            "type": "string"
          }
        },
        "required": ["name", "character", "profileImage"],
        "type": "object",
        "additionalProperties": false
      },
      "Video": {
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "fileSrc": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "runtime": {
            "type": "number",
            "format": "double"
          },
          "imgSrc": {
            "type": "string"
          },
          "imgUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "continueWatching": {
            "items": {
              "$ref": "#/components/schemas/ContinueWatching"
            },
            "type": "array"
          },
          "watchLists": {
            "items": {
              "$ref": "#/components/schemas/WatchList"
            },
            "type": "array"
          },
          "mediaInfo": {
            "$ref": "#/components/schemas/MediaInfo"
          },
          "videoTracks": {
            "items": {
              "$ref": "#/components/schemas/VideoTrack"
            },
            "type": "array"
          },
          "subtitleTracks": {
            "items": {
              "$ref": "#/components/schemas/SubtitleTrack"
            },
            "type": "array"
          },
          "audioTracks": {
            "items": {
              "$ref": "#/components/schemas/AudioTrack"
            },
            "type": "array"
          },
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/Chapter"
            },
            "type": "array"
          },
          "selectedAudioTrack": {
            "type": "number",
            "format": "double"
          },
          "selectedSubtitleTrack": {
            "type": "number",
            "format": "double"
          },
          "extraType": {
            "type": "string"
          },
          "episodeId": {
            "type": "string"
          },
          "movieId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "fileSrc",
          "hash",
          "runtime",
          "imgSrc",
          "imgUrls",
          "continueWatching",
          "watchLists"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "Episode": {
        "properties": {
          "id": {
            "type": "string"
          },
          "seasonId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nameLock": {
            "type": "boolean"
          },
          "year": {
            "type": "string"
          },
          "yearLock": {
            "type": "boolean"
          },
          "overview": {
            "type": "string"
          },
          "overviewLock": {
            "type": "boolean"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "directedBy": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "directedByLock": {
            "type": "boolean"
          },
          "writtenBy": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "writtenByLock": {
            "type": "boolean"
          },
          "episodeNumber": {
            "type": "number",
            "format": "double"
          },
          "seasonNumber": {
            "type": "number",
            "format": "double"
          },
          "order": {
            "type": "number",
            "format": "double"
          },
          "video": {
            "$ref": "#/components/schemas/Video"
          }
        },
        "required": [
          "id",
          "seasonId",
          "name",
          "nameLock",
          "year",
          "yearLock",
          "overview",
          "overviewLock",
          "score",
          "directedBy",
          "directedByLock",
          "writtenBy",
          "writtenByLock",
          "episodeNumber",
          "seasonNumber",
          "order",
          "video"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "WatchList": {
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "seriesId": {
            "type": "string"
          },
          "seasonId": {
            "type": "string"
          },
          "episodeId": {
            "type": "string"
          },
          "episode": {
            "$ref": "#/components/schemas/Episode"
          },
          "movieId": {
            "type": "string"
          },
          "movie": {
            "$ref": "#/components/schemas/Movie"
          },
          "videoId": {
            "type": "string"
          },
          "video": {
            "$ref": "#/components/schemas/Video"
          },
          "timeWatched": {
            "type": "number",
            "format": "double"
          },
          "watched": {
            "type": "boolean"
          },
          "lastWatched": {
            "type": "string"
          }
        },
        "required": ["id", "userId", "timeWatched", "watched", "lastWatched"],
        "type": "object",
        "additionalProperties": false
      },
      "Movie": {
        "properties": {
          "id": {
            "type": "string"
          },
          "libraryId": {
            "type": "string"
          },
          "imdbId": {
            "type": "string"
          },
          "themdbId": {
            "type": "number",
            "format": "double"
          },
          "imdbScore": {
            "type": "number",
            "format": "double"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "order": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "nameLock": {
            "type": "boolean"
          },
          "overview": {
            "type": "string"
          },
          "overviewLock": {
            "type": "boolean"
          },
          "year": {
            "type": "string"
          },
          "yearLock": {
            "type": "boolean"
          },
          "tagline": {
            "type": "string"
          },
          "taglineLock": {
            "type": "boolean"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "genresLock": {
            "type": "boolean"
          },
          "productionStudios": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "productionStudiosLock": {
            "type": "boolean"
          },
          "directedBy": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "directedByLock": {
            "type": "boolean"
          },
          "writtenBy": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "writtenByLock": {
            "type": "boolean"
          },
          "creator": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "creatorLock": {
            "type": "boolean"
          },
          "musicComposer": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "musicComposerLock": {
            "type": "boolean"
          },
          "cast": {
            "items": {
              "$ref": "#/components/schemas/CastData"
            },
            "type": "array"
          },
          "videoSrc": {
            "type": "string"
          },
          "musicSrc": {
            "type": "string"
          },
          "folder": {
            "type": "string"
          },
          "logoSrc": {
            "type": "string"
          },
          "logosUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "backgroundSrc": {
            "type": "string"
          },
          "backgroundsUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "coverSrc": {
            "type": "string"
          },
          "coversUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "watchLists": {
            "items": {
              "$ref": "#/components/schemas/WatchList"
            },
            "type": "array"
          },
          "videos": {
            "items": {
              "$ref": "#/components/schemas/Video"
            },
            "type": "array"
          },
          "extras": {
            "items": {
              "$ref": "#/components/schemas/Video"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "libraryId",
          "imdbId",
          "themdbId",
          "imdbScore",
          "score",
          "order",
          "name",
          "nameLock",
          "overview",
          "overviewLock",
          "year",
          "yearLock",
          "tagline",
          "taglineLock",
          "genres",
          "genresLock",
          "productionStudios",
          "productionStudiosLock",
          "directedBy",
          "directedByLock",
          "writtenBy",
          "writtenByLock",
          "creator",
          "creatorLock",
          "musicComposer",
          "musicComposerLock",
          "cast",
          "videoSrc",
          "musicSrc",
          "folder",
          "logoSrc",
          "logosUrls",
          "backgroundSrc",
          "backgroundsUrls",
          "coverSrc",
          "coversUrls",
          "watchLists",
          "videos",
          "extras"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "Chapter": {
        "properties": {
          "title": {
            "type": "string"
          },
          "time": {
            "type": "number",
            "format": "double"
          },
          "displayTime": {
            "type": "string"
          },
          "thumbnailSrc": {
            "type": "string"
          }
        },
        "required": ["title", "time", "displayTime", "thumbnailSrc"],
        "type": "object",
        "additionalProperties": false
      },
...
      "SetVideoWatchStateDTO": {
        "properties": {
          "watched": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": ["watched", "userId"],
        "type": "object",
        "additionalProperties": false
      },
...
      "ChangeIdentificationDTO": {
        "properties": {
          "themdbId": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["themdbId"],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_Movie_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Movie"
              }
            ],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "UpdateMovieDTO": {
        "properties": {
          "title": {
            "type": "string"
          },
          "releaseDate": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "SetMovieWatchStateDTO": {
        "properties": {
          "watched": {
            "type": "boolean"
          }
        },
        "required": ["watched"],
        "type": "object",
        "additionalProperties": false
      },
      "MovieResult": {
        "properties": {
          "poster_path": {
            "type": "string"
          },
          "adult": {
            "type": "boolean"
          },
          "overview": {
            "type": "string"
          },
          "release_date": {
            "type": "string"
          },
          "genre_ids": {
            "items": {
              "type": "number",
              "format": "double"
            },
            "type": "array"
          },
          "id": {
            "type": "number",
            "format": "double"
          },
          "media_type": {
            "type": "string",
            "enum": ["movie"],
            "nullable": false
          },
          "original_title": {
            "type": "string"
          },
          "original_language": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "backdrop_path": {
            "type": "string"
          },
          "popularity": {
            "type": "number",
            "format": "double"
          },
          "vote_count": {
            "type": "number",
            "format": "double"
          },
          "video": {
            "type": "boolean"
          },
          "vote_average": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["media_type"],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_MovieResult-Array_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/MovieResult"
            },
            "type": "array",
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_number_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
...
      "SetSeriesWatchStateDTO": {
        "properties": {
          "watched": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": ["watched", "userId"],
        "type": "object",
        "additionalProperties": false
      },
...
      "SetSeasonWatchStateDTO": {
        "properties": {
          "watched": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": ["watched", "userId"],
        "type": "object",
        "additionalProperties": false
      },
...
      "SetEpisodeWatchStateDTO": {
        "properties": {
          "state": {
            "type": "boolean"
          }
        },
        "required": ["state"],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Full user authentication with library and remote access checks"
      },
      "cookieAuthFast": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Fast user authentication without additional checks"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Admin user authentication"
      },
      "managementAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Management access for local or admin users"
      },
      "public": {
        "type": "apiKey",
        "in": "header",
        "name": "x-public-access",
        "description": "Public access without authentication"
      }
    }
  },
  "info": {
    "title": "Seerial Media Server API",
    "version": "0.2.0",
    "description": "Media management server API for the Seerial suite",
    "license": {
      "name": "GPL-3.0-only"
    },
    "contact": {
      "name": "Juan Llavero"
    }
  },
  "paths": {
...
    "/videos/{id}/watch-state": {
      "post": {
        "operationId": "SetWatchState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Set video watch state for a user",
        "tags": ["Videos"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetVideoWatchStateDTO"
              }
            }
          }
        }
      }
    },
...
    "/movies/{id}/metadata": {
      "post": {
        "operationId": "RefreshMovieMetadata",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Refresh movie metadata from external sources",
        "tags": ["Movies"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/movies/{id}/identification": {
      "post": {
        "operationId": "ChangeIdentification",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Change movie identification (TMDB ID)",
        "tags": ["Movies"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeIdentificationDTO"
              }
            }
          }
        }
      }
    },
    "/movies/{id}": {
      "patch": {
        "operationId": "Update",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Movie_"
                }
              }
            }
          }
        },
        "description": "Update movie details",
        "tags": ["Movies"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMovieDTO"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "Delete",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Delete a movie",
        "tags": ["Movies"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "get": {
        "operationId": "Get",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Movie_"
                }
              }
            }
          }
        },
        "description": "Get movie by ID",
        "tags": ["Movies"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/movies/{id}/watch-state": {
      "post": {
        "operationId": "SetWatchState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Set movie watch state for a user",
        "tags": ["Movies"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMovieWatchStateDTO"
              }
            }
          }
        }
      }
    },
    "/movies/search": {
      "get": {
        "operationId": "SearchMovies",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MovieResult-Array_"
                }
              }
            }
          }
        },
        "description": "Search movies in TMDB",
        "tags": ["Movies"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/movies/imdb-score": {
      "get": {
        "operationId": "GetImdbScore",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_number_"
                }
              }
            }
          }
        },
        "description": "Get IMDB score for a movie",
        "tags": ["Movies"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/movies/{id}/remaining-videos": {
      "get": {
        "operationId": "GetRemainingVideos",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_number_"
                }
              }
            }
          }
        },
        "description": "Get remaining videos count for a movie",
        "tags": ["Movies"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/series/{id}/watch-state": {
      "post": {
        "operationId": "SetWatchState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Set series watch state for a user",
        "tags": ["Series"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSeriesWatchStateDTO"
              }
            }
          }
        }
      }
    },
...
    "/seasons/{id}/watch-state": {
      "post": {
        "operationId": "SetWatchState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Set season watch state for a user",
        "tags": ["Seasons"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSeasonWatchStateDTO"
              }
            }
          }
        }
      }
    },
...
    "/episodes/{id}/watch-state": {
      "post": {
        "operationId": "SetWatchState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Set episode watch state for a user",
        "tags": ["Episodes"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetEpisodeWatchStateDTO"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:34200/api"
    }
  ]
}